% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{compute_outside_share}
\alias{compute_outside_share}
\title{Compute Share of Outside Good}
\usage{
compute_outside_share(df, mkt_share, market_id)
}
\arguments{
\item{df}{Dataframe to work with}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{market_id}{List of variables contaning market identifiers}
}
\value{
A dataframe with the variable \code{outside_share} added as a column.
}
\description{
Compute Share of Outside Good
}
\examples{
# Add Outside Share column to data
mkts <- create_markets(geog_id = "market_id", time_id = "year")
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- compute_outside_share(df, mkt_share = 'mkt_share', market_id = mkts)
}
