% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_demand.R
\name{create_equation}
\alias{create_equation}
\title{Create Regression Equation for Demand Estimation}
\usage{
create_equation(market_ids, product_id, market_share, outside_share,
  exog_charac, price, nest_shares, marketFE, productFE, instruments = NULL)
}
\arguments{
\item{market_ids}{List of variables contaning market identifiers}

\item{market_share}{Variable containing product market shares, as a character string}

\item{outside_share}{Variable containing outside market shares, as a character string}

\item{exog_charac}{Vector of exogenous product characteristic variables, as a character vector}

\item{price}{Variable containing market price, as a character string}

\item{nest_shares}{Vector of nest share variables, as a character vector}

\item{marketFE}{Character string of which market fixed effects to contain,
takes value "both", "geog" or "time"}

\item{instruments}{Vector of instrument names, as a character vector}
}
\value{
A regression formula.
}
\description{
Create Regression Equation for Demand Estimation
}
