% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_demand.R
\name{create_fe}
\alias{create_fe}
\title{Create Partial Formula for Fixed Effects}
\usage{
create_fe(market_ids, product_id, market_fe = "both", product_fe = TRUE)
}
\arguments{
\item{market_ids}{List of variables contaning market identifiers}

\item{market_fe}{Character string of which market fixed effects to contain,}
}
\value{
A character string with a partial formula for market fixed effetcs.
}
\description{
Create Partial Formula for Fixed Effects
}
