% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_demand.R
\name{create_lhs}
\alias{create_lhs}
\title{Create Left Hand Side of Regression Equation}
\usage{
create_lhs(exog_charac, price, nest_shares)
}
\arguments{
\item{exog_charac}{Vector of exogenous product characteristic variables, as a character vector}

\item{price}{Variable containing market price, as a character string}

\item{nest_shares}{Vector of nest share variables, as a character vector}
}
\value{
A character string with a partial formula of product characteristics and nest shares.
}
\description{
Create Left Hand Side of Regression Equation
}
