% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markets.R
\name{create_markets}
\alias{create_markets}
\title{Create list of market identifiers.}
\usage{
create_markets(geog_id, time_id)
}
\arguments{
\item{geog_id}{Variable that represents geographical market identifier as a string.}

\item{time_id}{Variable that represents time identifier as a string.}
}
\value{
A list containing \code{geog_id} and \code{time_id}.
}
\description{
Create list of market identifiers.
}
\examples{
# cross section data
create_markets(geog_id = "state", time_id = NULL)
# time series data
create_markets(geog_id = NULL, time_id = "year")
# panel data
create_markets(geog_id = "state", time_id = "year")
}
