% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_demand.R
\name{create_rhs}
\alias{create_rhs}
\title{Create Regression Equation for Demand Estimation}
\usage{
create_rhs(mkt_share, out_share)
}
\arguments{
\item{mkt_share}{Variable containing product market shares, as a character string}

\item{out_share}{Variable containing outside market shares, as a character string}
}
\value{
A character string with name of dependent variable.
}
\description{
Create Regression Equation for Demand Estimation
}
