% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{create_shares}
\alias{create_shares}
\title{Compute Share of Outside Good and Nest Shares}
\usage{
create_shares(df, market_id, mkt_share, outside_share = NULL,
  nest_id = NULL, subnest_id = NULL)
}
\arguments{
\item{df}{Dataframe to work with}

\item{market_id}{List of variables contaning market identifiers}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{outside_share}{Variable outside shares, as a character string.}

\item{nest_id}{Variable contaning first level nest, as a character string.}

\item{subnest_id}{Variable contaning second level nest, as a character string.}
}
\value{
A dataframe with the outside- and nest-shares added.
}
\description{
Compute Share of Outside Good and Nest Shares
}
\examples{
# Add Outside Share and Nest Share for two Level Nest
mkts <- create_markets(geog_id = "market_id", time_id = "year")
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- create_shares(df, market_id  = mkts,
                          mkt_share  = 'mkt_share',
                          nest_id    = 'class_id',
                          subnest_id = 'domestic')
# Add Outside Share and Nest Share for One Level Nest
mkts <- create_markets(geog_id = "market_id", time_id = "year")
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- create_shares(df, market_id  = mkts,
                          mkt_share  = 'mkt_share',
                          nest_id    = 'class_id')
# Add Nest Share for two Level Nest when outside share already computed
mkts <- create_markets(geog_id = "market_id", time_id = "year")
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- compute_outside_share(df, mkt_share = 'mkt_share', market_id = mkts)
df3  <- create_shares(df2, market_id    = mkts,
                           mkt_share     = 'mkt_share',
                           outside_share = 'outside_share',
                           nest_id       = 'class_id')
}
