% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_demand.R
\name{estimate_demand}
\alias{estimate_demand}
\title{Estimate a Demand Model From Aggregate Market Shares}
\usage{
estimate_demand(df, market_ids, market_share, outside_share, exog_charac, price,
  nest_shares = NULL, instruments = NULL, marketFE = "both",
  supply_side = FALSE)
}
\arguments{
\item{df}{Dataframe to containing data to be used in estimation}

\item{market_ids}{List of variables contaning market identifiers}

\item{market_share}{Variable containing product market shares, as a character string}

\item{outside_share}{Variable outside shares, as a character string}

\item{exog_charac}{Vector of exogenous product characteristic variables, as a character vector}

\item{price}{Variable containing market price, as a character string}

\item{nest_shares}{Vector of nest share variables, as a character vector}

\item{instruments}{Vector of instrument names, as a character vector - NOT CURRENTLY IMPLEMENTED}

\item{marketFE}{Character string of which market fixed effects to contain,
takes value "both", "geog" or "time"}

\item{supply_side}{Logical variable, \code{TRUE} leads to estimation with the
constraints from a Bertrand-Nash Oligopoly Model - NOT CURRENTLY IMPLEMENTED}
}
\value{
Returns an object of class "lm" when instruments = FALSE
}
\description{
Estimate a Demand Model From Aggregate Market Shares
}
\examples{
# Regression With Two Level Nest
mkts <- create_markets(geog_id = "market_id", time_id = "year")
price_var     <- c("price_euro")
exog_var      <- c("kilowatts", "fueleff_avg", "width", "height", "domestic")
mkt_share     <- c("mkt_share")
outside_share <- c("outside_share")
## Create Shares
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- create_shares(df, market_id  = mkts,
                          mkt_share  = 'mkt_share')
## Run regression
reg <- estimate_demand(df             = df2,
                        market_ids    = mkts,
                        market_share  = mkt_share,
                        outside_share = outside_share,
                        exog_charac   = exog_var,
                        price         = price_var,
                        marketFE      = "both")

# Regression With One Level Nest
mkts <- create_markets(geog_id = "market_id", time_id = "year")
price_var     <- c("price_euro")
exog_var      <- c("kilowatts", "fueleff_avg", "width", "height", "domestic")
mkt_share     <- c("mkt_share")
outside_share <- c("outside_share")
## Create Shares
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- create_shares(df, market_id  = mkts,
                          mkt_share  = 'mkt_share',
                          nest_id    = 'class_id')
## Run regression
reg <- estimate_demand(df             = df2,
                        market_ids    = mkts,
                        market_share  = mkt_share,
                        outside_share = outside_share,
                        exog_charac   = exog_var,
                        price         = price_var,
                        nest_shares   = c("within_nest"),
                        marketFE      = "both")

# Regression With Two Level Nest
mkts <- create_markets(geog_id = "market_id", time_id = "year")
price_var     <- c("price_euro")
exog_var      <- c("kilowatts", "fueleff_avg", "width", "height", "domestic")
mkt_share     <- c("mkt_share")
outside_share <- c("outside_share")
## Create Shares
df   <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
df2  <- create_shares(df, market_id  = mkts,
                          mkt_share  = 'mkt_share',
                          nest_id    = 'class_id',
                          subnest_id = 'domestic')
## Run regression
reg <- estimate_demand(df             = df2,
                        market_ids    = mkts,
                        market_share  = mkt_share,
                        outside_share = outside_share,
                        exog_charac   = exog_var,
                        price         = price_var,
                        nest_shares   = c("within_nest", "within_subnest"),
                        marketFE      = "both")
}
