% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-eurocars.R
\docType{data}
\name{eurocars}
\alias{eurocars}
\title{eurocars: Annual Sales of Cars in Europe}
\format{A data.frame with 11,549 observations on 39 variables:
\itemize{
 \item \strong{To be Done:} Numeric identifier for model of aircraft
}}
\usage{
data('eurocars')
}
\description{
Annual model-level sales of cars in European countries between 1970 and 1999.
}
\examples{
 str(eurocars)
}
\keyword{datasets}
