% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{gen_subnest_share}
\alias{gen_subnest_share}
\title{Calculate Within Shares for Subnest Structure}
\usage{
gen_subnest_share(df, mkt_share, market_id, nest_id, subnest_id)
}
\arguments{
\item{df}{Dataframe to work with}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{nest_id}{Variable containing market share of nest, as a character string.}

\item{subnest_id}{Variable containing market share of subnest, as a character string.}
}
\value{
A dataframe with the outside- and nest-shares added.
}
\description{
Calculate Within Shares for Subnest Structure
}
