% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{gen_within_share}
\alias{gen_within_share}
\title{Calculate Within Shares for Nesting Structure}
\usage{
gen_within_share(df, mkt_share, nest_share, subnest_share = NULL)
}
\arguments{
\item{df}{Dataframe to work with}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{nest_share}{Variable containing market share of nest, as a character string.}

\item{subnest_share}{Variable containing market share of subnest, as a character string.}
}
\value{
A dataframe with the outside- and nest-shares added.
}
\description{
Calculate Within Shares for Nesting Structure
}
