% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jets.R
\docType{data}
\name{jets}
\alias{jets}
\title{jets: Annual Sales of Corporate Jets}
\format{A data.frame with 56 observations on 9 variables:
\itemize{
 \item \strong{model_id:} Numeric identifier for model of aircraft
 \item \strong{manufacturer_id:} Numeric identifier for manufacturer of aircraft
 \item \strong{year:} Year of the observation
 \item \strong{price:} Price in millions of US dollars
 \item \strong{quantity:} Units sold
 \item \strong{max_range:} Range in miles
 \item \strong{speed:} Speed in miles per hour
 \item \strong{cabin_volume:} Cabin Space in cubic feet
 \item \strong{brand:} Manufacturer
}}
\usage{
data('jets')
}
\description{
Annual model-level aircraft sales in the large-cabin segment of the businessjet market between 1985 and 1998.
}
\examples{
 str(jets)
}
\keyword{datasets}
