% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{mkt_share_from_sales}
\alias{mkt_share_from_sales}
\title{Create Market Shares from Quantity Sold}
\usage{
mkt_share_from_sales(df, quantity, population, frac_pop = 1)
}
\arguments{
\item{df}{Dataframe to work with}

\item{quantity}{Variable containing number of goods sold, as a character string.}

\item{population}{Variable containing number of number of individuals
in the market, as a string.}

\item{frac_pop}{Fraction of population to use when obtaining
market shares, between 0 and 1.}
}
\value{
A dataframe with the variable \code{mkt_share} added as a column.
}
\description{
Create Market Shares from Quantity Sold
}
\examples{
# Relevant market is entire population
new_df <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population')
# Relevant market is 1/2 the entire population
new_df <- mkt_share_from_sales(eurocars,
                               quantity   = 'qty_sold',
                               population = 'population',
                               frac_pop   =  0.5)
}
