% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{within_share_one_nest}
\alias{within_share_one_nest}
\title{Calculate Product level Within Shares for One Layer Nest}
\usage{
within_share_one_nest(df, mkt_share, nest_share)
}
\arguments{
\item{df}{Dataframe to work with}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{nest_share}{Variable containing market share of nest, as a character string.}
}
\value{
A dataframe with the outside- and nest-shares added.
}
\description{
Calculate Product level Within Shares for One Layer Nest
}
