% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shares.R
\name{within_share_two_nest}
\alias{within_share_two_nest}
\title{Calculate Product level Within Shares for Two Layer Nest}
\usage{
within_share_two_nest(df, mkt_share, nest_share, subnest_share)
}
\arguments{
\item{df}{Dataframe to work with}

\item{mkt_share}{Variable containing product market shares, as a character string.}

\item{nest_share}{Variable containing market share of nest, as a character string.}

\item{subnest_share}{Variable containing market share of subnest, as a character string.}
}
\value{
A dataframe with the outside- and nest-shares added.
}
\description{
Calculate Product level Within Shares for Two Layer Nest
}
